chore: restructure Domain layer to screaming architecture
