+++ backend/src/SmartHome.Domain/Device/Thermostat/ThermostatMode.cs
@@ -1,4 +1,4 @@
-namespace SmartHome.Domain.Enum;
+namespace SmartHome.Domain.Device.Thermostat;
 
 public enum ThermostatMode
 {
