+++ backend/src/SmartHome.Infrastructure/Persistence/SmartHomeDbContext.cs
@@ -1,4 +1,8 @@
 using Microsoft.EntityFrameworkCore;
+using SmartHome.Domain.Device.Light;
+using SmartHome.Domain.Device.Fan;
+using SmartHome.Domain.Device.Thermostat;
+using SmartHome.Domain.Device.DoorLock;
 using SmartHome.Domain.Device;
 
 namespace SmartHome.Infrastructure.Persistence;
